Adds a new views file , controllers and models
