+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "server.js",
   "scripts": {
+    "seed": "node scripts/seed.js",
     "test": "npm run lint",
     "start": "node server.js",
     "lint": "eslint --quiet .",
@@ -15,12 +16,14 @@
   "license": "ISC",
   "dependencies": {
     "bcryptjs": "2.4.3",
-    "express": "^4.17.0",
+    "express": "^4.17.1",
+    "express-handlebars": "^5.1.0",
     "express-session": "^1.16.1",
-    "mysql2": "^1.6.5",
-    "passport": "^0.4.0",
+    "mysql2": "^1.7.0",
+    "passport": "^0.4.1",
     "passport-local": "^1.0.0",
-    "sequelize": "^5.8.6"
+    "sequelize": "^5.22.3",
+    "sequelize-cli": "^5.5.1"
   },
   "devDependencies": {
     "chai": "^4.1.2",
