+++ server.js
@@ -1,36 +1,33 @@
-// Requiring necessary npm packages
 const express = require("express");
-const session = require("express-session");
-// Requiring passport as we've configured it
-const passport = require("./config/passport");
-
-// Setting up port and requiring models for syncing
-const PORT = process.env.PORT || 8080;
 const db = require("./models");
+const exphbs = require("express-handlebars");
+const cocktailcontroller = require("./controllers/cocktails");
+const htmlcontroller = require("./controllers/html");
 
-// Creating express app and configuring middleware needed for authentication
+// Sets up the Express App
+// =============================================================
 const app = express();
+const PORT = process.env.PORT || 8080;
+
+// Sets up the Express app to handle data parsing
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
+
+app.engine("handlebars", exphbs({ defaultLayout: "main" }));
+app.set("view engine", "handlebars");
+// Static directory
 app.use(express.static("public"));
-// We need to use sessions to keep track of our user's login status
-app.use(
-  session({ secret: "keyboard cat", resave: true, saveUninitialized: true })
-);
-app.use(passport.initialize());
-app.use(passport.session());
 
-// Requiring our routes
-require("./routes/html-routes.js")(app);
-require("./routes/api-routes.js")(app);
+// Routes
+// =============================================================
+require("./controllers/cocktails.js")(app);
+require("./controllers/html.js")(app);
+cocktailcontroller(app);
+htmlcontroller(app);
 
-// Syncing our database and logging a message to the user upon success
-db.sequelize.sync().then(() => {
-  app.listen(PORT, () => {
-    console.log(
-      "==> 🌎  Listening on port %s. Visit http://localhost:%s/ in your browser.",
-      PORT,
-      PORT
-    );
-  });
+db.sequelize.sync();
+// Starting our Express app
+// =============================================================
+app.listen(PORT, () => {
+  console.log(`App listening on PORT ${PORT}`);
 });
