+++ views/cocktails.handlebars
@@ -1,12 +1,16 @@
-  
-<h1>Cocktails</h1>
+
+ <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <!--Import materialize.css-->
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
+    <link rel="stylesheet" href="style.css">
+    <h1 class="center">Cocktails</h1>
 {{#each cocktails}}
-<h2>{{ name }}</h2>
-<img src="{{ imageUrl }}" alt="{{ name }}">
-<h3>Recipe</h3>
-<p>{{ recipe }}</p>
-<h3>Ingredients</h3>
-<ul>
+<h2 class="center">{{ name }}</h2>
+<img class="circle center"src="{{ imageUrl }}" alt="{{ name }}">
+<h3 class="center" >Recipe</h3>
+<p class="center flow-text">{{ recipe }}</p>
+<h3 class="center">Ingredients</h3>
+<ul class="center flow-text">
   {{#each Ingredients}}
   <li>{{ name }}</li>
   {{/each}}
@@ -14,3 +18,6 @@
 {{else}}
 <p>No matching cocktails found.</p>
 {{/each}}
+<script src="https://code.jquery.com/jquery-3.5.1.min.js"
+    integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
