+++ views/home.handlebars
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <!--Import materialize.css-->
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
+    <link rel="stylesheet" href="style.css">
+    <!--Let browser know website is optimized for mobile-->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <h1 class="center"> MIX IT UP </h1>
+    <nav>
+    <div class="nav-wrapper">
+      <a href="#" class="brand-logo center"><i class="material-icons">local_bar</i></a>
+      <ul id="nav-mobile" class="left hide-on-med-and-down">
+        <li><a href="/cocktails">View All Cocktails</a></li>
+        
+      </ul>
+    </div>
+  </nav>
+  <div class="center">
+  <h3 class="center flow-text"> An application that gives you  access to cocktail recipes based on inspiration from ingredients that are already available in your home !   
+ </h3>
+  <a class="btn-floating btn-large waves-effect waves-light center green z-depth-4" href="/mix"><i class="material-icons">local_bar</i></a>
+</div>
+</body>
+<script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
+</html>
