+++ views/mix.handlebars
@@ -1,0 +1,110 @@
+<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+<!--Import materialize.css-->
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
+<link rel="stylesheet" href="style.css">
+
+
+<h2 class="center">Choose your Alcohol!</h2>
+
+<form class="center" action="#">
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Vodka</span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Gin</span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input class="with-gap" name="group1" type="radio" />
+            <span>Brandy</span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Whiskey</span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Tequila</span>
+        </label>
+    </p>
+</form>
+
+<h2 class="center">Choose your sweetner!</h2>
+
+<form class="center" action="#">
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Fruit Juice</span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Simple Syrup</span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input class="with-gap" name="group1" type="radio" />
+            <span>Agave</span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Fruit</span>
+        </label>
+    </p>
+
+</form>
+
+<h2 class="center">Choose your Sour!</h2>
+
+<form class="center" action="#">
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Lemon</span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Lime </span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input class="with-gap" name="group1" type="radio" />
+            <span>Grapefruit</span>
+        </label>
+    </p>
+    <p>
+        <label>
+            <input name="group1" type="radio" class="with-gap" />
+            <span>Other</span>
+        </label>
+    </p>
+  <a class="btn-floating btn-large waves-effect waves-light center orange darken-4 z-depth-4" href="/cocktails"><i class="material-icons">local_bar</i></a>
+   
+</form>
+
+
+</form>
+
+  
+</div>
+<script src="https://code.jquery.com/jquery-3.5.1.min.js"
+    integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
